See README for changelogs

+++ frontend/src/routes/app/servers/[id]/+page.server.ts
@@ -4,11 +4,12 @@ import type { PageServerLoad, Actions } from './$types';
 import { error, redirect, fail } from '@sveltejs/kit';
 import { findServerMonitoringByServerId } from '$lib/server/db/monitoring';
 import { setMailing } from '$lib/server/db/mailing';
+import { hasPermission, PERMISSIONS } from '@shared/roles';
 
 export const load = (async ({ params, locals: { user } }) => {
   const { id } = params;
 
-  if(!user) {
+  if (!user) {
     throw error(401, 'Unauthorized');
   }
 
@@ -18,14 +19,19 @@ export const load = (async ({ params, locals: { user } }) => {
       throw error(404, 'Server not found');
     }
     const monitoring = await findServerMonitoringByServerId(parseInt(id));
-    return { server: server.server, mailingEnabled: server.mailingEnabled , monitoring };
+    return { server: server.server, mailingEnabled: server.mailingEnabled, monitoring };
   } catch (err: unknown) {
+    console.log(err);
     throw error(500, err as string);
   }
 }) satisfies PageServerLoad;
 
 export const actions: Actions = {
-  deleteServer: async ({ params: { id } }) => {
+  deleteServer: async ({ params: { id }, locals: { user } }) => {
+    if (!hasPermission(user?.role, PERMISSIONS.DELETE_SERVER)) {
+      return fail(403, { error: true, message: 'Forbidden' });
+    }
+
     try {
       const success = await deleteServer(parseInt(id));
 
@@ -57,24 +63,32 @@ export const actions: Actions = {
     if (!user) {
       return fail(401, { error: true, message: 'Unauthorized' });
     }
+    if (!hasPermission(user.role, PERMISSIONS.UPDATE_SERVER)) {
+      return fail(403, { error: true, message: 'Forbidden' });
+    }
     const formData = Object.fromEntries(await request.formData());
     const {
       deleteServerNameInput: name,
       deleteServerDescriptionInput: description,
       deleteServerPublicUrlInput: publicUrl,
-      emailAlerts
+      emailAlerts,
+      monitored
     } = formData as {
       deleteServerNameInput: string;
       deleteServerDescriptionInput: string;
       deleteServerPublicUrlInput: string;
-      emailAlerts: string
+      emailAlerts: string;
+      monitored: string;
     };
+
+    const isMonitored = monitored === 'on';
 
     try {
       const status = await updateServer(parseInt(id), {
         name,
         description,
-        publicUrl
+        publicUrl,
+        monitored: isMonitored
       });
 
       if (!status) {
@@ -86,7 +100,6 @@ export const actions: Actions = {
       if (!emailingStatus) {
         return fail(500, { error: true, message: 'Failed to update server' });
       }
-
 
       return { success: true, message: 'Server updated' };
     } catch (error) {
